Fix optimal counter strategy
The issue is I was returning the EV from one node in an information set before
the entire information set was comptued. That EV wasn't always based on the
move that we finally decided from the information set. To fix this I changed it
to a two pass BFS thing.